Breaking up files again since it's really fast to create them too big..
